misc: janitorial changes, minor bug fix
BUG FIX: in the engine, there was an issue when inserting a null
`votingEndTime` into `swapArray`.

I moved that logic into the "view lobby: voting phase"
logic since this is the only place where we use the variable, and it
guarantees that the voting end time will not be null.